test : 출석 확인 기능 관련 단위테스트 추가
