+++ src/test/java/test/model/administration/AttendanceStatisticsTest.java
@@ -7,9 +7,4 @@ import org.junit.jupiter.api.Test;
 
 public class AttendanceStatisticsTest {
 
-    @DisplayName("")
-    @Test
-    void test() {
-        assertThat();
-    }
 }
