+++ src/test/java/test/model/attendance/AttendanceStatisticTest.java
@@ -1,13 +1,14 @@
 package test.model.attendance;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import model.attendance.AttendanceStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 public class AttendanceStatisticTest {
 
-    @DisplayName("")
-    @Test
-    void test() {
-        assertThat();
-    }
+
 }
