+++ src/test/java/test/model/attendance/CrewTest.java
@@ -5,9 +5,5 @@ import org.junit.jupiter.api.Test;
 
 public class CrewTest {
 
-    @DisplayName("")
-    @Test
-    void test() {
-        assertThat();
-    }
+
 }
